doc: Add a new "Object hierarchy" section.
This new section in documentation page can be used to display object
hierarchy, including inheritance relations.

This commit is a basic implementation used to show object inheritance
in the context of the `wibox.widget.textclock` widget. This is an
interesting example because textclock are actually based on textbox.
Thanks to this section, it makes it clear of the user when reading doc.